cr feedback, move new test into its own class
